fix: adjust entities attributes - standardize types

+++ src/main/java/com/example/marisa/model/entities/Cashier.java
@@ -3,39 +3,39 @@ package com.example.marisa.model.entities;
 import com.example.marisa.model.enumeration.CashierStatusEnum;
 
 public class Cashier {
-    private int id;
-    private float openingBalance;
-    private float finalBalance;
+    private Integer id;
+    private double openingBalance;
+    private double finalBalance;
     private CashierStatusEnum status;
 
-    public Cashier(int id, float openingBalance, float finalBalance, String status) {
+    public Cashier(Integer id, double openingBalance, double finalBalance, String status) {
         this.id = id;
         this.openingBalance = openingBalance;
         this.finalBalance = finalBalance;
         this.status = CashierStatusEnum.valueOf(status);
     }
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public float getOpeningBalance() {
+    public double getOpeningBalance() {
         return openingBalance;
     }
 
-    public void setOpeningBalance(float openingBalance) {
+    public void setOpeningBalance(double openingBalance) {
         this.openingBalance = openingBalance;
     }
 
-    public float getFinalBalance() {
+    public double getFinalBalance() {
         return finalBalance;
     }
 
-    public void setFinalBalance(float finalBalance) {
+    public void setFinalBalance(double finalBalance) {
         this.finalBalance = finalBalance;
     }
